Refactor input handling for wind speed and solar radiation to use integer type for improved validation

+++ src/main.py
@@ -110,12 +110,12 @@ while True:
                     case "3":
                         try:
                             telemetria_configurada = True
-                            colonia.vento = float(input("Velocidade do vento (m/s) [0 a 30]: "))
+                            colonia.vento = int(input("Velocidade do vento (m/s) [0 a 30]: "))
                             while colonia.vento < 0 or colonia.vento > 30:
-                                colonia.vento = float(input("Velocidade do vento (m/s) [0 a 30]: "))
-                            colonia.radiacao_solar = float(input("Radiação Solar (%) [0 a 100]: "))
+                                colonia.vento = int(input("Velocidade do vento (m/s) [0 a 30]: "))
+                            colonia.radiacao_solar = int(input("Radiação Solar (%) [0 a 100]: "))
                             while colonia.radiacao_solar < 0 or colonia.radiacao_solar > 100:
-                                colonia.radiacao_solar = float(input("Radiação Solar (%) [0 a 100]: "))
+                                colonia.radiacao_solar = int(input("Radiação Solar (%) [0 a 100]: "))
                             print("Vento e Radiação Solar Atualizados!")
                         except ValueError:
                             telemetria_configurada = False
